+++ src/restfw_admin/CHANGES.rst
@@ -14,6 +14,11 @@
 CHANGELOG
 *********
 
+Next release
+============
+
+- Fixed errors with nullable schema nodes.
+
 0.2 (2020-08-06)
 ================